Split out BlitzEvaluator from RoF Rename skills to actions
