+++ src/parser/jobs/mnk/modules/Forms.tsx
@@ -12,7 +12,7 @@ import {Data} from 'parser/core/modules/Data'
 import Downtime from 'parser/core/modules/Downtime'
 import Suggestions, {SEVERITY, Suggestion, TieredSuggestion} from 'parser/core/modules/Suggestions'
 import React from 'react'
-import {FORM_TIMEOUT_MILLIS, FORMS, OPO_OPO_SKILLS} from './constants'
+import {FORM_TIMEOUT_MILLIS, FORMS, OPO_OPO_ACTIONS} from './constants'
 import {fillActions, fillStatuses} from './utilities'
 
 export class Forms extends Analyser {
@@ -39,7 +39,7 @@ export class Forms extends Analyser {
 
 	override initialise(): void {
 		this.forms = fillStatuses(FORMS, this.data)
-		this.opoOpoSkills = fillActions(OPO_OPO_SKILLS, this.data)
+		this.opoOpoSkills = fillActions(OPO_OPO_ACTIONS, this.data)
 
 		const playerFilter = filter<Event>().source(this.parser.actor.id)
 		this.addEventHook(playerFilter.type('statusApply').status(oneOf(this.forms)), this.onGain)
