+++ src/parser/jobs/mnk/modules/constants.ts
@@ -10,32 +10,32 @@ export const FORMS: StatusKey[] = [
 	'COEURL_FORM',
 ]
 
-export const OPO_OPO_SKILLS: ActionKey[] = [
+export const OPO_OPO_ACTIONS: ActionKey[] = [
 	'BOOTSHINE',
 	'DRAGON_KICK',
 	'SHADOW_OF_THE_DESTROYER',
 ]
 
-export const RAPTOR_SKILLS: ActionKey[] = [
+export const RAPTOR_ACTIONS: ActionKey[] = [
 	'TRUE_STRIKE',
 	'TWIN_SNAKES',
 	'FOUR_POINT_FURY',
 ]
 
-export const COEURL_SKILLS: ActionKey[] = [
+export const COEURL_ACTIONS: ActionKey[] = [
 	'SNAP_PUNCH',
 	'DEMOLISH',
 	'ROCKBREAKER',
 ]
 
-export const FORM_SKILLS: ActionKey[] = [
-	...OPO_OPO_SKILLS,
-	...RAPTOR_SKILLS,
-	...COEURL_SKILLS,
+export const FORM_ACTIONS: ActionKey[] = [
+	...OPO_OPO_ACTIONS,
+	...RAPTOR_ACTIONS,
+	...COEURL_ACTIONS,
 ]
 
 // Blitzes
-export const BLITZ_SKILLS: ActionKey[] = [
+export const BLITZ_ACTIONS: ActionKey[] = [
 	'ELIXIR_FIELD',
 	'CELESTIAL_REVOLUTION',
 	'RISING_PHOENIX',
